Change Kbps representation from INT to DOUBLE

+++ backend/NetboardApi/Dtos/TrafficWindowDtos/CreateTrafficWindowDto.cs
@@ -4,7 +4,7 @@ public sealed record CreateTrafficWindowDto
 {
     public string DeviceIp { get; init; } = string.Empty;
     public string ProtocolName { get; init; } = string.Empty;
-    public int TotalSizeKbps { get; init; }
-    public int UploadSizeKbps { get; init; }
-    public int DownloadSizeKbps { get; init; }
+    public double TotalSizeKbps { get; init; }
+    public double UploadSizeKbps { get; init; }
+    public double DownloadSizeKbps { get; init; }
 }
